Windows: port NCCL to compile and run on Windows/WDDM
Build system:
- CMakeLists.txt: add Windows platform detection, MSVC flags, DLL
  export target, conditional linking (no ibverbs/mlx5 on Windows),
  Ninja generator support; add unit-test and demo targets
- build.bat: Windows one-click CMake configure + Ninja build script
- src/misc/gen_nccl_h.py: generate nccl.h on Windows (no sed)
- src/nccl.def: DLL export map for MSVC linker

Platform stubs and OS abstraction:
- src/os/windows.cc / windows_stubs.cc: implement POSIX-like APIs
  (pthread, semaphore, mmap, clock_gettime, sysconf, popen, etc.)
- src/include/dlfcn_win.h: dlopen/dlsym shim via LoadLibrary
- src/include/nccl_win_pch.h: Windows precompiled-header guard
- src/include/compiler/msvc.h: MSVC attribute / intrinsic compat
- src/os/linux.cc: guard Linux-only includes

MSVC / CCCL compatibility throughout the codebase:
- Replace Linux-only headers (unistd.h, sys/mman.h, etc.) with
  Windows equivalents or guards in: alloc.h, bitops.h, collectives.h,
  core.h, debug.h, gdrwrap.h, graph.h, ipcsocket.h, nccl_common.h,
  ras.h, socket.h, os.h, nccl.h.in, and all plugin/* wrappers
- Fix __attribute__ → __declspec and packed-struct attributes in
  gin_proxy_device_host_common.h, gin_gdaki*.h, nccl_device/utility.h
- Guard ibverbs / mlx5dv / RAS / GDR symbols with NCCL_OS_LINUX
- Guard nvtx3 C++ API with Windows-safe include order (nvtxImpl.h)
- Fix CCCL/Thrust headers for MSVC in device/prims_ll*.h,
  reduce_kernel.h, device/generate.py, device/symmetric/generate.py

Runtime portability:
- src/allocator.cc, init.cc, proxy.cc, mem_manager.cc, bootstrap.cc,
  transport.cc, dev_runtime.cc, debug.cc, graph/topo.cc,
  graph/tuning.cc, graph/xml.{cc,h}: guard Linux-only ioctls,
  fork/exec, NUMA, proc-filesystem reads, and ibverbs calls
- src/graph/topo.cc: tolerate missing "arch" attribute in topology XML
  (Windows topology generator does not emit it); default to x86_64
- src/include/device.h: set ncclCollUnroll=4 for SM_120 (workstation
  Blackwell has only 53 KB shared memory; unroll=8 overflows)
- src/misc/shmutils.cc (calloc NULL-check), param.cc, socket.cc,
  utils.cc, misc/*wrap.cc: Windows-safe implementations
- src/transport/net.cc, net_socket.cc: guard Linux-only socket opts
- src/rma/rma_proxy.cc, scheduler/allgatherv_sched.cc,
  gin/gin_host_proxy.{cc,h}, ras/ras.cc: guard Linux-only paths

Tests and samples:
- tests/unit/nccl_tests.cu: Windows-compatible test harness
- tests/demo/nccl_demo.cpp, samples/nccl_demo.cu: portable demo apps

+++ src/os/linux.cc
@@ -92,7 +92,7 @@ static ncclResult_t setCpuStackSize() {
   return ncclSuccess;
 }
 
-extern int ncclParamSetCpuStackSize();
+extern int64_t ncclParamSetCpuStackSize();
 
 ncclResult_t ncclOsInitialize() {
   if (ncclParamSetCpuStackSize() != 0) {
@@ -109,8 +109,8 @@ bool ncclOsSocketIsValid(struct ncclSocket* sock) {
   return ncclOsSocketDescriptorIsValid(sock->socketDescriptor);
 }
 
-extern int ncclParamPollTimeOut();
-extern int ncclParamRetryTimeOut();
+extern int64_t ncclParamPollTimeOut();
+extern int64_t ncclParamRetryTimeOut();
 
 void ncclOsPollSocket(int socketDescriptor, int op) {
   struct pollfd pfd;
@@ -120,7 +120,7 @@ void ncclOsPollSocket(int socketDescriptor, int op) {
   (void) poll(&pfd, 1, ncclParamPollTimeOut());
 }
 
-extern long int ncclParamRetryCnt();
+extern int64_t ncclParamRetryCnt();
 
 ncclResult_t ncclOsSocketTryAccept(struct ncclSocket* sock) {
   socklen_t socklen = sizeof(union ncclSocketAddress);
@@ -144,8 +144,8 @@ ncclResult_t ncclOsSocketTryAccept(struct ncclSocket* sock) {
   return ncclSuccess;
 }
 
-extern int ncclParamSocketMaxRecvBuff();
-extern int ncclParamSocketMaxSendBuff();
+extern int64_t ncclParamSocketMaxRecvBuff();
+extern int64_t ncclParamSocketMaxSendBuff();
 
 ncclResult_t ncclOsSocketSetFlags(struct ncclSocket* sock) {
   ncclResult_t ret = ncclSuccess;
